Fix two warnings in gz_physics-main-cnlwin (#734)
(cherry picked from commit 4f7e2941b816bb3ae25403632e508de4044111e1)

+++ test/integration/DoublePendulum.cc
@@ -113,7 +113,7 @@ void DoublePendulum_TEST(ignition::plugin::PluginPtr _plugin)
   ignition::math::PID pid0(100, 0, 10);
   ignition::math::PID pid1(10, 0, 5);
   const std::chrono::duration<double> settleTime(std::chrono::seconds(4));
-  unsigned int settleSteps = settleTime / dt;
+  unsigned int settleSteps = static_cast<unsigned int>(settleTime / dt);
   for (unsigned int i = 0; i < settleSteps; ++i)
   {
     auto positions = output.Get<JointPositions>();
